requirements.txt
cython==0.20.1
nuitka==0.5.1.1

M2Crypto==0.22.3
pycrypto==2.6.1
